ctf-engine/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
httpx==0.27.0
pydantic==2.7.4
python-dotenv==1.0.1
docker==7.1.0
aiofiles==23.2.1
requests==2.32.3
litellm>=1.80.0
jinja2>=3.1.0
tavily-python>=0.5.0
